# src/public/openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "Digital Cart API",
    "version": "1.0.0"
  },
  "servers": [
    {
      "url": "http://localhost:8080/",
      "description": "Ambiente local"
    }
  ],
  "tags": [
    { "name": "Usuários" },
    { "name": "Transações" }
  ],
  "components": {
    "securitySchemes": {
      "sanctum": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "Token"
      }
    }
  },
  "security": [
    { "sanctum": [] }
  ],
  "paths": {
    "/users": {
      "post": {
        "security": [],
        "tags": ["Usuários"],
        "summary": "Criar usuário",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": ["name", "email", "password"],
                "properties": {
                  "name": { "type": "string" },
                  "email": { "type": "string", "format": "email" },
                  "password": { "type": "string" }
                }
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Usuário criado",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": { "type": "boolean", "example": true },
                    "code": { "type": "integer", "example": 201 },
                    "data": {
                      "type": "object",
                      "properties": {
                        "user": { "type": "object" },
                        "token": { "type": "string" }
                      }
                    },
                    "error": { "nullable": true }
                  }
                }
              }
            }
          },
          "422": {
            "description": "Erro de validação",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": { "type": "boolean", "example": false },
                    "code": { "type": "integer", "example": 422 },
                    "data": { "type": "object", "nullable": true },
                    "error": {
                      "type": "object",
                      "properties": {
                        "message": { "type": "string", "example": "Erro de validação" },
                        "errors": { "type": "object" }
                      }
                    }
                  }
                }
              }
            }
          },
          "500": {
            "description": "Erro interno",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": { "type": "boolean", "example": false },
                    "code": { "type": "integer", "example": 500 },
                    "data": { "type": "object", "nullable": true },
                    "error": { "type": "string", "example": "Erro ao criar usuário" }
                  }
                }
              }
            }
          }
        }
      }
    },
    "/users/balance": {
      "get": {
        "tags": ["Usuários"],
        "summary": "Consultar saldo",
        "responses": {
          "200": {
            "description": "Saldo atual",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": { "type": "boolean", "example": true },
                    "code": { "type": "integer", "example": 200 },
                    "data": {
                      "type": "object",
                      "properties": {
                        "balance": { "type": "number", "format": "float" }
                      }
                    },
                    "error": { "type": "object", "nullable": true }
                  }
                }
              }
            }
          },
          "401": { "description": "Não autenticado" },
          "500": { "description": "Erro interno" }
        }
      }
    },
    "/deposit": {
      "post": {
        "tags": ["Transações"],
        "summary": "Depositar",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": ["amount"],
                "properties": {
                  "amount": { "type": "number", "format": "float", "minimum": 0.01 }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Depósito realizado",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": { "type": "boolean", "example": true },
                    "code": { "type": "integer", "example": 200 },
                    "data": { "type": "object", "properties": { "message": { "type": "string" } } },
                    "error": { "type": "object", "nullable": true }
                  }
                }
              }
            }
          },
          "422": {
            "description": "Erro de validação",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": { "type": "boolean", "example": false },
                    "code": { "type": "integer", "example": 422 },
                    "data": { "type": "object", "nullable": true },
                    "error": { "type": "object", "properties": { "message": { "type": "string" }, "errors": { "type": "object" } } }
                  }
                }
              }
            }
          },
          "401": {
            "description": "Não autenticado",
            "content": {
              "application/json": {
                "schema": { "type": "object", "properties": { "success": { "type": "boolean", "example": false }, "code": { "type": "integer", "example": 401 }, "data": { "type": "object", "nullable": true }, "error": { "type": "string" } } }
              }
            }
          },
          "500": {
            "description": "Erro interno",
            "content": {
              "application/json": {
                "schema": { "type": "object", "properties": { "success": { "type": "boolean", "example": false }, "code": { "type": "integer", "example": 500 }, "data": { "type": "object", "nullable": true }, "error": { "type": "string" } } }
              }
            }
          }
        }
      }
    },
    "/withdraw": {
      "post": {
        "tags": ["Transações"],
        "summary": "Sacar",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": ["amount"],
                "properties": {
                  "amount": { "type": "number", "format": "float", "minimum": 0.01 }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Saque realizado",
            "content": { "application/json": { "schema": { "type": "object", "properties": { "success": { "type": "boolean", "example": true }, "code": { "type": "integer", "example": 200 }, "data": { "type": "object", "properties": { "message": { "type": "string" } } }, "error": { "type": "object", "nullable": true } } } } }
          },
          "422": {
            "description": "Erro de validação",
            "content": { "application/json": { "schema": { "type": "object", "properties": { "success": { "type": "boolean", "example": false }, "code": { "type": "integer", "example": 422 }, "data": { "type": "object", "nullable": true }, "error": { "type": "object", "properties": { "message": { "type": "string" }, "errors": { "type": "object" } } } } } } }
          },
          "401": {
            "description": "Não autenticado",
            "content": { "application/json": { "schema": { "type": "object", "properties": { "success": { "type": "boolean", "example": false }, "code": { "type": "integer", "example": 401 }, "data": { "type": "object", "nullable": true }, "error": { "type": "string" } } } } }
          },
          "500": {
            "description": "Erro interno",
            "content": { "application/json": { "schema": { "type": "object", "properties": { "success": { "type": "boolean", "example": false }, "code": { "type": "integer", "example": 500 }, "data": { "type": "object", "nullable": true }, "error": { "type": "string" } } } } }
          }
        }
      }
    },
    "/transfer": {
      "post": {
        "tags": ["Transações"],
        "summary": "Transferir",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": ["amount", "recipient"],
                "properties": {
                  "amount": { "type": "number", "format": "float", "minimum": 0.01 },
                  "recipient": { "type": "string", "format": "email" }
                }
              }
            }
          }
        },
        "responses": {
          "200": { "description": "Transferência realizada", "content": { "application/json": { "schema": { "type": "object", "properties": { "success": { "type": "boolean", "example": true }, "code": { "type": "integer", "example": 200 }, "data": { "type": "object", "properties": { "message": { "type": "string" } } }, "error": { "type": "object", "nullable": true } } } } } },
          "422": { "description": "Erro de validação", "content": { "application/json": { "schema": { "type": "object", "properties": { "success": { "type": "boolean", "example": false }, "code": { "type": "integer", "example": 422 }, "data": { "type": "object", "nullable": true }, "error": { "type": "object", "properties": { "message": { "type": "string" }, "errors": { "type": "object" } } } } } } } },
          "401": { "description": "Não autenticado", "content": { "application/json": { "schema": { "type": "object", "properties": { "success": { "type": "boolean", "example": false }, "code": { "type": "integer", "example": 401 }, "data": { "type": "object", "nullable": true }, "error": { "type": "string" } } } } } },
          "500": { "description": "Erro interno", "content": { "application/json": { "schema": { "type": "object", "properties": { "success": { "type": "boolean", "example": false }, "code": { "type": "integer", "example": 500 }, "data": { "type": "object", "nullable": true }, "error": { "type": "string" } } } } } }
        }
      }
    },
    "/history": {
      "get": {
        "tags": ["Transações"],
        "summary": "Histórico de transações",
        "responses": {
          "200": {
            "description": "Lista de transações",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": { "type": "boolean", "example": true },
                    "code": { "type": "integer", "example": 200 },
                    "data": {
                      "type": "object",
                      "properties": {
                        "transactions": {
                          "type": "array",
                          "items": {
                            "type": "object",
                            "properties": {
                              "id": { "type": "integer" },
                              "type": { "type": "string" },
                              "created_at": { "type": "string", "format": "date-time" },
                              "amount": { "type": "number", "format": "float" },
                              "recipient": { "type": "string", "nullable": true },
                              "sender": { "type": "string", "nullable": true }
                            }
                          }
                        }
                      }
                    },
                    "error": { "type": "object", "nullable": true }
                  }
                }
              }
            }
          },
          "401": { "description": "Não autenticado" },
          "500": { "description": "Erro interno" }
        }
      }
    }
  }
}